DownloadLogsCommandBaseHandler
