+++ src/Octoshift/Commands/DownloadLogsCommandBase.cs
@@ -13,12 +13,7 @@ namespace OctoshiftCLI.Commands;
 
 public class DownloadLogsCommandBase : Command
 {
-    private readonly OctoLogger _log;
-    private readonly ITargetGithubApiFactory _githubApiFactory;
-    private readonly HttpDownloadService _httpDownloadService;
-    private readonly RetryPolicy _retryPolicy;
-
-    internal Func<string, bool> FileExists = path => File.Exists(path);
+    protected DownloadLogsCommandBaseHandler BaseHandler { get; init; }
 
     public DownloadLogsCommandBase(
         OctoLogger log,
@@ -26,10 +21,7 @@ public class DownloadLogsCommandBase : Command
         HttpDownloadService httpDownloadService,
         RetryPolicy retryPolicy) : base(name: "download-logs", description: "Downloads migration logs for migrations.")
     {
-        _log = log;
-        _githubApiFactory = githubApiFactory;
-        _httpDownloadService = httpDownloadService;
-        _retryPolicy = retryPolicy;
+        BaseHandler = new DownloadLogsCommandBaseHandler(log, githubApiFactory, httpDownloadService, retryPolicy);
     }
 
     protected virtual Option<string> GithubOrg { get; } = new("--github-org")
@@ -83,71 +75,6 @@ public class DownloadLogsCommandBase : Command
         AddOption(MigrationLogFile);
         AddOption(Overwrite);
         AddOption(Verbose);
-    }
-
-    public async Task Handle(DownloadLogsCommandArgs args)
-    {
-        if (args is null)
-        {
-            throw new ArgumentNullException(nameof(args));
-        }
-
-        _log.Verbose = args.Verbose;
-
-        _log.LogWarning("Migration logs are only available for 24 hours after a migration finishes!");
-
-        _log.LogInformation("Downloading migration logs...");
-        _log.LogInformation($"{GithubOrg.GetLogFriendlyName()}: {args.GithubOrg}");
-        _log.LogInformation($"{GithubRepo.GetLogFriendlyName()}: {args.GithubRepo}");
-
-        if (args.GithubApiUrl.HasValue())
-        {
-            _log.LogInformation($"{GithubApiUrl.GetLogFriendlyName()}: {args.GithubApiUrl}");
-        }
-
-        if (args.GithubPat.HasValue())
-        {
-            _log.LogInformation($"{GithubPat.GetLogFriendlyName()}: ***");
-        }
-
-        if (args.MigrationLogFile.HasValue())
-        {
-            _log.LogInformation($"{MigrationLogFile.GetLogFriendlyName()}: {args.MigrationLogFile}");
-        }
-
-        args.MigrationLogFile ??= $"migration-log-{args.GithubOrg}-{args.GithubRepo}.log";
-
-        if (FileExists(args.MigrationLogFile))
-        {
-            if (!args.Overwrite)
-            {
-                throw new OctoshiftCliException($"File {args.MigrationLogFile} already exists!  Use --overwrite to overwrite this file.");
-            }
-
-            _log.LogWarning($"Overwriting {args.MigrationLogFile} due to --overwrite option.");
-        }
-
-        var githubApi = _githubApiFactory.Create(args.GithubApiUrl, args.GithubPat);
-
-        var result = await _retryPolicy.RetryOnResult(async () => await githubApi.GetMigrationLogUrl(args.GithubOrg, args.GithubRepo), string.Empty,
-            "Waiting for migration log to populate...");
-
-        if (result.Outcome == OutcomeType.Successful && result.Result is null)
-        {
-            throw new OctoshiftCliException($"Migration for repository {args.GithubRepo} not found!");
-        }
-
-        if (result.Outcome == OutcomeType.Failure)
-        {
-            throw new OctoshiftCliException($"Migration log for repository {args.GithubRepo} unavailable!");
-        }
-
-        var logUrl = result.Result;
-
-        _log.LogInformation($"Downloading log for repository {args.GithubRepo} to {args.MigrationLogFile}...");
-        await _httpDownloadService.Download(logUrl, args.MigrationLogFile);
-
-        _log.LogSuccess($"Downloaded {args.GithubRepo} log to {args.MigrationLogFile}.");
     }
 }
