+++ src/OctoshiftCLI.Tests/Octoshift/Commands/DownloadLogsCommandBaseTests.cs
@@ -9,7 +9,7 @@ namespace OctoshiftCLI.Tests.Octoshift.Commands;
 
 public class DownloadLogsCommandBaseTests
 {
-    private readonly DownloadLogsCommandBase _command;
+    private readonly DownloadLogsCommandBaseHandler _command;
     private readonly Mock<GithubApi> _mockGithubApi = TestHelpers.CreateMock<GithubApi>();
     private readonly Mock<ITargetGithubApiFactory> _mockGithubApiFactory = new();
     private readonly Mock<HttpDownloadService> _mockHttpDownloadService = TestHelpers.CreateMock<HttpDownloadService>();
@@ -17,7 +17,7 @@ public class DownloadLogsCommandBaseTests
 
     public DownloadLogsCommandBaseTests()
     {
-        _command = new DownloadLogsCommandBase(
+        _command = new DownloadLogsCommandBaseHandler(
             _mockLogger.Object,
             _mockGithubApiFactory.Object,
             _mockHttpDownloadService.Object,
